Move Engine signature AST traversals/transforms into `./transforms` module.
These AST visitors and transformations are more generally usable for other
purposes rather than just the Apollo Engine signature reporting and would
seem to belong in a module of their own.
